actually move stuff back to hw_clip lmao
