+++ src/hardware/r_opengl/r_opengl.h
@@ -123,6 +123,8 @@ extern GLint			maximumAnisotropy;
 extern INT32            oglflags;
 extern GLint            textureformatGL;
 
+extern GLfloat projMatrix[16];
+
 typedef enum
 {
 	GLF_NOZBUFREAD = 0x01,
